Refactor code into views and utils and fix imports
