+++ ap_src/ap_app/utils/objects.py
@@ -1,4 +1,4 @@
-from .models import UserProfile
+from ..models import UserProfile
 
 def find_user_obj(user_to_find):
     """Finds & Returns object of 'UserProfile' for a user
