+++ ap_src/ap_app/views/views.py
@@ -13,13 +13,13 @@ from django.shortcuts import redirect, render
 from django.urls import reverse_lazy
 from django.views.generic import CreateView
 
-from .objects import obj_exists, find_user_obj
+from ..utils.objects import obj_exists, find_user_obj
 from .calendarview import main_calendar
 
-from .forms import AcceptPolicyForm, AbsenceForm, DeleteUserForm
-from .models import Absence, UserProfile, ColourScheme, ColorData, RecurringException
+from ..forms import AcceptPolicyForm, AbsenceForm, DeleteUserForm
+from ..models import Absence, UserProfile, ColourScheme, ColorData, RecurringException
 
-from .utils.switch_permissions import check_for_lingering_switch_perms, remove_switch_permissions
+from ..utils.switch_permissions import check_for_lingering_switch_perms, remove_switch_permissions
 
 User = get_user_model()
